refactor(perf): split /ai-ops + /youtube-strategy into server shell + motion island
13th + 14th in the pattern series.
